+++ astropy/units/quantity.py
@@ -1386,10 +1386,16 @@ class Quantity(np.ndarray, metaclass=InheritDocstrings):
         try:
             _value = value.to_value(self.unit)
         except AttributeError:
+            # We're not a Quantity, so let's try a more general conversion.
+            # Plain arrays will be converted to dimensionless in the process,
+            # but anything with a unit attribute will use that.
             try:
-                _value = dimensionless_unscaled.to(self.unit, value)
+                _value = Quantity(value).to_value(self.unit)
             except UnitsError as exc:
-                if can_have_arbitrary_unit(value):
+                # last chance: if this was not something with a unit
+                # and is all 0, inf, or nan, we treat it as arbitrary unit.
+                if (not hasattr(value, 'unit') and
+                        can_have_arbitrary_unit(value)):
                     _value = value
                 else:
                     raise exc